fix[funcoes]: fiz pequenos consertos nas funcoes de busca, baixa de estadia e calculo de pontos de fidelidade

+++ Main/Funcoes/BaixaDeEstadia/baixaDeEstadia.h
@@ -8,6 +8,6 @@
 #include <vector>
 
 
-float baixaDeEstadia(int code);
+int baixaDeEstadia(int code);
 
 #endif
